requirements.txt
python-dotenv~=1.0.0
Pyrogram~=2.0.106
TgCrypto~=1.2.5
